pyproject.toml
[project]
name = "snwflk-dbt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-core>=1.10.13",
    "dbt-snowflake>=1.10.2",
    "httpx>=0.28.1",
    "invoke>=2.2.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "snowflake>=1.8.0",
    "streamlit>=1.50.0",
]

[dependency-groups]
dev = [
    "icecream>=2.1.8",
]
